Conversion cleanup :)

+++ src/hotspot/share/gc/z/zAdaptiveHeap.cpp
@@ -509,10 +509,10 @@ uint64_t ZAdaptiveHeap::soft_ref_delay() {
   const double avg_time_since_last = _old_data._gc_times_since_last.avg();
   const size_t old_live = stats->live_at_mark_end();
   const size_t old_used = ZHeap::heap()->used_old();
-  const size_t old_allocated = old_used - old_live;
-  const double old_alloc_rate = MAX2((double)old_allocated / M / avg_time_since_last, 1.0);
+  const double old_allocated = old_used - old_live;
+  const double old_alloc_rate = MAX2(old_allocated / M / avg_time_since_last, 1.0);
 
-  const double time_to_old_oom = (double)free_heap / M / old_alloc_rate;
+  const double time_to_old_oom = double(free_heap) / M / old_alloc_rate;
 
   const double free_ratio = double(target_capacity) / double(free_heap);
